orchestrator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pyjwt==2.8.0
boto3==1.29.7
sqlalchemy==2.0.23
pydantic==2.5.0
websockets==12.0
slowapi==0.1.9
email-validator==2.1.0
hvac==2.1.0
